UPDATE added jSymbolic features, moved file operations out of AudioFileConversion
