+++ amuse/src/amuse/util/converters/ConverterInterface.java
@@ -1,14 +1,9 @@
 package amuse.util.converters;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
-import java.nio.channels.FileChannel;
 import java.util.ArrayList;
 import java.util.List;
-
-import org.apache.log4j.Level;
 
 import amuse.data.io.ArffDataSet;
 import amuse.data.io.DataSetAbstract;
@@ -16,8 +11,6 @@ import amuse.data.io.attributes.Attribute;
 import amuse.data.modality.Format;
 import amuse.interfaces.nodes.NodeException;
 import amuse.preferences.AmusePreferences;
-import amuse.util.AmuseLogger;
-import amuse.util.audio.AudioFileConversion;
 
 public interface ConverterInterface {
 	
@@ -39,48 +32,6 @@ public interface ConverterInterface {
 		}
 		return targetFormats;
 	}
-	
-	public static void deleteFile(File wavFile) {
-        boolean success = wavFile.delete();
-        if (!success) {
-        	wavFile.deleteOnExit();
-        }
-    }
-	
-	public static void fileCopy(File srcFile, File destFile) {
-    	FileInputStream srcChannelFIS = null;
-    	FileOutputStream dstChannelFOS = null;
-        try {
-            // Create channel on the source
-        	srcChannelFIS = new FileInputStream(srcFile);
-            FileChannel srcChannel = srcChannelFIS.getChannel();
-
-            // Create channel on the destination
-            dstChannelFOS = new FileOutputStream(destFile);
-            FileChannel dstChannel = dstChannelFOS.getChannel();
-
-            // Copy file contents from source to destination
-            dstChannel.transferFrom(srcChannel, 0, srcChannel.size());
-
-            // Close the channels
-            srcChannel.close();
-            dstChannel.close();
-        } catch (IOException e) {
-            AmuseLogger.write(AudioFileConversion.class.getName(), Level.ERROR, "Unable to copy " + srcFile.getAbsolutePath() + " to " + destFile.getAbsolutePath() + ".");
-        }
-        finally{
-        	if(srcChannelFIS != null){
-        		try {
-					srcChannelFIS.close();
-				} catch (IOException e) {}
-        	}
-        	if(dstChannelFOS != null){
-        		try {
-        			dstChannelFOS.close();
-				} catch (IOException e) {}
-        	}
-        }
-    }
 	
 	public static ConverterInterface getConversionClass(Format sourceFormat, Format targetFormat) {
 		DataSetAbstract configTable;
@@ -107,11 +58,6 @@ public interface ConverterInterface {
 			e.printStackTrace();
 		}
 		return null;
-	}
-	
-	public static String cutExtension(String filename) {
-		 int pos = filename.lastIndexOf('.');
-         return filename.substring(0, pos);
 	}
 	
 	public void convert(File file, File outputFolder) throws IOException, NodeException;
